python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bonsai-sdk"
version = "0.1.0"
description = "Python SDK for the Bonsai Network State Engine gRPC API"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",
    "pyyaml>=6.0",
    "paramiko>=3.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-cov", "requests"]
ml  = [
    "scikit-learn>=1.4",
    "joblib>=1.3",
    "pyarrow>=15.0",
    "pandas>=2.0",
    "numpy>=1.26",
]

[tool.hatch.build.targets.wheel]
packages = ["bonsai_sdk"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: requires live infrastructure (ContainerLab, Docker services, bonsai binary) — skipped in unit CI",
]

[tool.coverage.run]
source = ["../scripts"]
omit = ["*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
]
